[Twitch] Reworking entire codebase for multiple twitch connections
So the entire codebase was rewritten so that we could stop connecting to
a single twitch chat and instead connect to zero or more twitch chats.
Some optimizations were done to prevent the same bot account from
entering a chat more than once (not that this actually matters).
Probably needs some clean up. However, we can officially apply
subsections of our code to individual chats.